Fix lint errors after refactor
