+++ routes/memory_routes.py
@@ -1,0 +1,126 @@
+"""Memory-related API routes for the dashboard."""
+from flask import jsonify, request
+from datetime import datetime
+from zoneinfo import ZoneInfo
+import gc
+import psutil
+from collections import deque
+import sys
+import time
+
+from config import get_timezone
+from memory_utils import memory_usage_history, memory_usage_lock, log_memory_usage
+
+
+def register_memory_routes(app, state_manager, get_active_connections, server_start_time):
+    """Register memory management API routes."""
+
+    @app.route("/api/memory-profile")
+    def memory_profile():
+        """API endpoint for detailed memory profiling."""
+        try:
+            process = psutil.Process()
+            mem_info = process.memory_info()
+
+            type_counts = {}
+            for obj in gc.get_objects():
+                obj_type = type(obj).__name__
+                type_counts[obj_type] = type_counts.get(obj_type, 0) + 1
+
+            most_common = sorted(type_counts.items(), key=lambda x: x[1], reverse=True)[:15]
+
+            memory_trend = {}
+            if memory_usage_history:
+                recent = memory_usage_history[-1]
+                oldest = memory_usage_history[0] if len(memory_usage_history) > 1 else recent
+                memory_trend = {
+                    "oldest_timestamp": oldest.get("timestamp"),
+                    "recent_timestamp": recent.get("timestamp"),
+                    "growth_mb": recent.get("rss_mb", 0) - oldest.get("rss_mb", 0),
+                    "growth_percent": recent.get("percent", 0) - oldest.get("percent", 0),
+                }
+
+            return jsonify(
+                {
+                    "memory": {
+                        "rss_mb": mem_info.rss / 1024 / 1024,
+                        "vms_mb": mem_info.vms / 1024 / 1024,
+                        "percent": process.memory_percent(),
+                        "data_structures": {
+                            "arrow_history": {
+                                "entries": sum(
+                                    len(v)
+                                    for v in state_manager.get_history().values()
+                                    if isinstance(v, (list, deque))
+                                ),
+                                "keys": list(state_manager.get_history().keys()),
+                            },
+                            "metrics_log": {"entries": len(state_manager.get_metrics_log())},
+                            "memory_usage_history": {"entries": len(memory_usage_history)},
+                            "sse_connections": get_active_connections(),
+                        },
+                        "most_common_objects": dict(most_common),
+                        "trend": memory_trend,
+                    },
+                    "gc": {
+                        "garbage": len(gc.garbage),
+                        "counts": gc.get_count(),
+                        "threshold": gc.get_threshold(),
+                        "enabled": gc.isenabled(),
+                    },
+                    "system": {
+                        "uptime_seconds": (datetime.now(ZoneInfo(get_timezone())) - server_start_time).total_seconds(),
+                        "python_version": sys.version,
+                    },
+                }
+            )
+        except Exception as exc:  # pragma: no cover
+            import logging
+
+            logging.error("Error in memory profiling: %s", exc)
+            return jsonify({"error": str(exc)}), 500
+
+    @app.route("/api/memory-history")
+    def memory_history_api():
+        """API endpoint for memory usage history."""
+        with memory_usage_lock:
+            history_copy = list(memory_usage_history)
+        return jsonify(
+            {
+                "history": history_copy,
+                "current": {
+                    "rss_mb": psutil.Process().memory_info().rss / 1024 / 1024,
+                    "percent": psutil.Process().memory_percent(),
+                },
+            }
+        )
+
+    @app.route("/api/force-gc", methods=["POST"])
+    def force_gc_api():
+        """API endpoint to force garbage collection."""
+        try:
+            generation = request.json.get("generation", 2) if request.is_json else 2
+            if generation not in [0, 1, 2]:
+                generation = 2
+
+            start_time = time.time()
+            objects_before = len(gc.get_objects())
+            collected = gc.collect(generation)
+            duration = time.time() - start_time
+            objects_after = len(gc.get_objects())
+            log_memory_usage(state_manager, get_active_connections())
+            return jsonify(
+                {
+                    "status": "success",
+                    "collected": collected,
+                    "duration_seconds": duration,
+                    "objects_removed": objects_before - objects_after,
+                    "generation": generation,
+                }
+            )
+        except Exception as exc:  # pragma: no cover
+            import logging
+
+            logging.error("Error during forced GC: %s", exc)
+            return jsonify({"status": "error", "message": str(exc)}), 500
+
